requirements.txt
asgiref==3.6.0
Django==4.1.7 
djangorestframework==3.14.0
protobuf==3.20.3
psycopg2-binary==2.9.5
sqlparse==0.4.3
